+++ app/src/androidTest/java/io/github/viabachelora23michaelkutaibakasper/bprapp/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
 package io.github.viabachelora23michaelkutaibakasper.bprapp
 
-import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
-
+import androidx.test.platform.app.InstrumentationRegistry
+import org.junit.Assert.*
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
