+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/data/sign_in/AuthenticationClient.kt
@@ -6,23 +6,15 @@ import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.activity.result.ActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.stringResource
-
 import com.google.android.gms.auth.api.signin.GoogleSignIn
-import com.google.android.gms.auth.api.signin.GoogleSignInAccount
-import com.google.android.gms.auth.api.signin.GoogleSignInClient
-import com.google.android.gms.auth.api.signin.GoogleSignInOptions
 import com.google.android.gms.common.api.ApiException
 import com.google.firebase.Firebase
-import com.google.firebase.auth.AuthCredential
 import com.google.firebase.auth.AuthResult
 import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.auth.GoogleAuthProvider
 import com.google.firebase.auth.auth
-import io.github.viabachelora23michaelkutaibakasper.bprapp.R
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
 
@@ -35,7 +27,6 @@ class AuthenticationClient : IAuthenticationClient {
         onAuthError: (ApiException) -> Unit
     ): ManagedActivityResultLauncher<Intent, ActivityResult> {
         val scope = rememberCoroutineScope()
-        val context = LocalContext.current.applicationContext
 
         return rememberLauncherForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
             Firebase.auth.signOut()
