+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/ui/navigation/Routes.kt
@@ -1,0 +1,19 @@
+package io.github.viabachelora23michaelkutaibakasper.bprapp.ui.navigation
+
+enum class BottomNavigationScreens() {
+    Map,
+    Recommendations,
+    Achievements,
+    Profile,
+    EventDetails
+}
+
+enum class CreateEventScreens() {
+    Title,
+    Location,
+    DateAndTime,
+    Details,
+    Images,
+    InviteFriends,
+    EventSummary
+}
